gather bytecode tests
