+++ zkevm-circuits/src/bytecode_circuit.rs
@@ -4,7 +4,7 @@
 pub mod bytecode_unroller;
 /// Bytecode circuit
 pub mod circuit;
+pub(crate) mod param;
 /// Bytecode circuit tester
 #[cfg(any(feature = "test", test))]
-pub mod dev;
-pub(crate) mod param;
+pub mod test;
