--- zkevm-circuits/src/bytecode_circuit/dev.rs
@@ -1,47 +1,0 @@
-use super::bytecode_unroller::{unroll, UnrolledBytecode};
-use super::circuit::BytecodeCircuit;
-
-use eth_types::Field;
-
-use halo2_proofs::dev::MockProver;
-use log::error;
-
-impl<F: Field> BytecodeCircuit<F> {
-    /// Verify that the selected bytecode fulfills the circuit
-    pub fn verify_raw(k: u32, bytecodes: Vec<Vec<u8>>) {
-        let unrolled: Vec<_> = bytecodes.iter().map(|b| unroll(b.clone())).collect();
-        Self::verify(k, unrolled, true);
-    }
-
-    pub(crate) fn verify(k: u32, bytecodes: Vec<UnrolledBytecode<F>>, success: bool) {
-        let circuit = BytecodeCircuit::<F>::new(bytecodes, 2usize.pow(k));
-
-        let prover = MockProver::<F>::run(k, &circuit, Vec::new()).unwrap();
-        let result = prover.verify();
-        if let Err(failures) = &result {
-            for failure in failures.iter() {
-                error!("{}", failure);
-            }
-        }
-        assert_eq!(result.is_ok(), success);
-    }
-}
-
-/// Test bytecode circuit with unrolled bytecode
-pub fn test_bytecode_circuit_unrolled<F: Field>(
-    k: u32,
-    bytecodes: Vec<UnrolledBytecode<F>>,
-    success: bool,
-) {
-    let circuit = BytecodeCircuit::<F>::new(bytecodes, 2usize.pow(k));
-
-    let prover = MockProver::<F>::run(k, &circuit, Vec::new()).unwrap();
-    let result = prover.verify_par();
-    if let Err(failures) = &result {
-        for failure in failures.iter() {
-            error!("{}", failure);
-        }
-    }
-    let error_msg = if success { "valid" } else { "invalid" };
-    assert_eq!(result.is_ok(), success, "proof must be {}", error_msg);
-}
